test: split accountability suites by behavior
